% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2excel.R
\name{extract2excel}
\alias{extract2excel}
\title{filter and save to excel file}
\usage{
extract2excel(.data, .title = NULL, .author = NULL, .word = NULL)
}
\arguments{
\item{.data}{tangshi}

\item{.title}{title}

\item{.author}{author name}

\item{.word}{keyword}
}
\description{
filter tangshi to new dataframe and save to excel file
}
\examples{
extract2excel(tangshi, .author = "白居易")

}
