% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tangshi}
\alias{tangshi}
\title{Data from poetry of the Tang Dynasty}
\format{
A data frame with 42945 rows and 5 variables:
\describe{
\item{block}{juan}
\item{id}{id}
\item{title}{title}
\item{author}{author}
\item{value}{value}
}
}
\usage{
tangshi
}
\description{
Data from poetry of the Tang Dynasty
}
\keyword{datasets}
