% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangshiR-package.R
\docType{package}
\name{tangshiR-package}
\alias{tangshiR}
\alias{tangshiR-package}
\title{tangshiR: Dataset for poetry of the Tang Dynasty}
\description{
data set of Tang shi.
}
\author{
\strong{Maintainer}: Minjie Wang \email{wangminjiede@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
